cleanup pt. 2: deleted more complicated messaging simulation
