+++ simulations/distributed/messaging_app.py
@@ -19,7 +19,9 @@ from pyspark.context import SparkContext
 from kafka import KafkaProducer
 
 from simulations.distributed.executor.spark.jobs import distributed_simulations
+from simulations.distributed.policies import enter_action, message_actions, exit_action
 from simulations.distributed.spark.session import sc
+from simulations.distributed.state_updates import send_message, count_messages, add_send_time, current_time
 
 
 def count(start, step):
@@ -36,91 +38,7 @@ intitial_conditions = {
     'total_send_time': 0.000000
 }
 
-
-# Actions
-def messages(client_id, room, action, _input, sender, receiver=None):
-    return {
-        'types': [action],
-        'messages': [
-            {
-                'client': client_id, 'room': room, 'action': action,
-                'sender': sender, 'receiver': receiver,
-                'input': _input,
-                'creatred': datetime.now()
-            }
-        ]
-    }
-
-
-def enter_action(state, room, user):
-    def f(_g, step, sL, s, kafkaConfig):
-        msgs = messages(state, room, 'enter', f"{user} enters {room}", user)
-        msgs['send_times'] = [0.000000]
-        msgs['msg_counts'] = [len(msgs['messages'])]
-        return msgs
-    return f
-
-def message_actions(state, room, _input, sender, receiver):
-    msgs = messages(state, room, 'send', _input, sender, receiver)
-    msgs_list = msgs['messages']
-    def send_action(_g, step, sL, s, kafkaConfig):
-        start_time = datetime.now()
-        for msg in msgs_list:
-            producer: KafkaProducer = kafkaConfig['producer']
-            topic: str = kafkaConfig['send_topic']
-            encoded_msg = str(msg).encode('utf-8')
-            producer.send(topic, encoded_msg)
-        msgs['send_times'] = [(datetime.now() - start_time).total_seconds()]
-        msgs['msg_counts'] = [len(msgs_list)]
-        return msgs
-
-    return send_action
-
-def exit_action(state, room, user):
-    def f(_g, step, sL, s, kafkaConfig):
-        msgs = messages(state, room, 'exit', f"{user} exited {room}", user)
-        msgs_list = msgs['messages']
-        msgs['send_times'] = [0.000000]
-        msgs['msg_counts'] = [len(msgs_list)]
-        return msgs
-    return f
-
 # State Updates
-def update_users(users, actions, action_types=['send','enter','exit']):
-    users = deepcopy(users)
-    for action_type in action_types:
-        if action_type in actions['types']:
-            for msg in actions['messages']:
-                if msg['action'] == 'send' and action_type == 'send':
-                    continue
-                elif msg['action'] == 'enter' and action_type == 'enter':
-                    for user in msg['sender']:
-                        users.append(user) # register_entered
-                elif msg['action'] == 'exit' and action_type == 'exit':
-                    for user in msg['sender']:
-                        users.remove(user) # remove_exited
-    return users
-
-add = lambda a, b: a + b
-def count_messages(_g, step, sL, s, actions, kafkaConfig):
-    return 'total_msg_count', s['total_msg_count'] + reduce(add, actions['msg_counts'])
-
-def add_send_time(_g, step, sL, s, actions, kafkaConfig):
-    return 'total_send_time', s['total_send_time'] + reduce(add, actions['send_times'])
-
-def send_message(state):
-    return lambda _g, step, sL, s, actions, kafkaConfig: (
-        state,
-        {
-            'users': update_users(s[state]['users'], actions),
-            'messages': actions['messages'],
-            'msg_counts': reduce(add, actions['msg_counts']),
-            'send_times': reduce(add, actions['send_times'])
-        }
-    )
-
-def current_time(state):
-    return lambda _g, step, sL, s, actions, kafkaConfig: (state, datetime.now())
 
 sim_composition = [
     {
